WIP: Try to fix #3581, make Screenshots work with third party cookies disabled
This adds a second attempt to login to wantsauth logins, one that runs in sitehelper.js, and tries to get the cookie set on a request that appears to come from the content page itself.

Unfortunately this doesn't work, as the XMLHttpRequest acts like it comes from a moz-extension URL and not the page, so it's still treated as a third party cookie.